Added delete comment functionality, fixed the url to an article

+++ app/Http/Controllers/Admin/CommentsController.php
@@ -3,14 +3,13 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Contracts\CommentContract;
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
+use App\Http\Controllers\BaseController;
 
 use Illuminate\Validation\ValidationException;
 use Illuminate\Database\QueryException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
-class CommentsController extends Controller
+class CommentsController extends BaseController
 {
     protected $commentContract;
     protected $commentRepository;
@@ -19,15 +18,15 @@ class CommentsController extends Controller
      * CommentsController constructor.
      * @param CommentContract $commentContract -> Comment db table handler
      */
-    public function __construct(CommentContract $commentContract)
+    public function __construct(CommentContract $commentRepository)
     {
-        $this->commentRepository = $commentContract;
+        $this->commentRepository = $commentRepository;
     }
 
     /**
      * Get all the articles in the table and return the view with the data.
      * @param $request: Request -> Search param
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View|\Illuminate\Http\RedirectResponse
      */
     public function index()
     {
@@ -51,18 +50,31 @@ class CommentsController extends Controller
     /**
      * Handle the request to delete a comment based on the ID provided
      * @param $id -> Comment ID.
+     * @param $blog_id -> Blog ID.
      * @return \Illuminate\Http\RedirectResponse -> Redirection with the message of success or error
      */
-    public function delete($id)
+    public function delete($id, $blog_id)
     {
         try {
+            $user = auth()->user();
+
+            if (!$user->articles()->where('id', $blog_id)->exists()) {
+                return redirect()->back()->with('error', 'Unauthorized Access');
+            }
+
+            $comment = $this->commentRepository->findCommentById($id);
+            
+            if (!$comment || $comment->blog_id != $blog_id) {
+                return redirect()->back()->with('error', 'Comment not found or does not belong to the specified blog');
+            }
+
             $comment = $this->commentRepository->deleteComment($id);
 
             if (!$comment) {
-                return $this->responseRedirectBack('Error occurred while deleting the comment.', 'error', true, true);
+                return redirect()->back()->with('error', 'Error occurred while deleting the comment.');
             }
 
-            return $this->responseRedirect('admin.comment.index', 'The comment was deleted successfully', 'success', false, false);
+            return redirect()->back()->with('success', 'The comment was deleted successfully.');
         } catch (ValidationException $e) {
             return $this->responseRedirectBack('Validation error: ' . $e->getMessage(), 'error', true, true)->withInput();
         } catch (QueryException $e) {
